Made basicmath code equal to OpenISA's
This made basicmath stop working. This is probably related to
long double operations.

+++ mibench/automotive/basicmath/cubic.c
@@ -16,9 +16,9 @@ void SolveCubic(double  a,
                 int    *solutions,
                 double *x)
 {
-      double    a1 = b/a, a2 = c/a, a3 = d/a;
-      double    Q = (a1*a1 - 3.0*a2)/9.0;
-      double R = (2.0*a1*a1*a1 - 9.0*a1*a2 + 27.0*a3)/54.0;
+      long double    a1 = b/a, a2 = c/a, a3 = d/a;
+      long double    Q = (a1*a1 - 3.0*a2)/9.0;
+      long double R = (2.0*a1*a1*a1 - 9.0*a1*a2 + 27.0*a3)/54.0;
       double    R2_Q3 = R*R - Q*Q*Q;
 
       double    theta;
